Fix  Bug 607767 -  integer overflow in brasero_audio2cue_len_to_string
Thanks to Jonathan Matthew for spotting this

+++ plugins/audio2cue/burn-audio2cue.c
@@ -310,7 +310,7 @@ brasero_audio2cue_len_to_string (guint64 len)
 {
 	int sec;
 	int min;
-	int frame;
+	guint64 frame;
 
 	if (len >= 1000000000)
 		frame = (len % 1000000000) * 75;
